commneted first 2 functions
